Switch to using a config file for app config and logging
This switches the consumer to use the fedora-messaging configuration
file for config values, including log configuration. It also moves the
Azure consumer to an `azure` module since in the future we will run
tests against all the clouds Fedora supports.

It also sets up an example configuration file which can be used to run
the consumer with `fedora-messaging --conf
./fedora-messaging.toml.example consume`.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>