Add a Containerfile and usage instructions
This builds a container with the application and LISA, and includes
instructions on using it.

We still need to identify the exact permissions required to run this in
Azure so we can create a role with the appropriate permissions. After
that, we can update the instructions to avoid assigning the service
principal a "Contributor" role.

Additionally, it would be great to have a conversation with the LISA
folks about publishing releases to PyPI so we don't have to do the whole
git checkout thing.

Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>